fix templates and views

+++ app-Concesionaria/App_Concesionaria/usuario/views/usuarios_view.py
@@ -3,13 +3,11 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import user_passes_test
 from django.utils.decorators import method_decorator
 from django.contrib.auth.models import User
-
 from usuario.repositorios.usuarios_repositories import Usuario_Repository
 from usuario.forms import UserRegisterForm, UserUpdateForm
 
 repository = Usuario_Repository()
 
-# Decorador para permitir acceso solo a usuarios staff
 def staff_required(user):
     return user.is_staff
 
@@ -31,20 +29,21 @@ class Usuario_Create(View):
         form = self.form_class(request.POST)
         if form.is_valid():
             user = form.save()
-            return redirect('usuarios_list')
+            return redirect('usuario_list')
         return render(request, 'usuarios/create.html', {'form': form})
 
 @method_decorator(user_passes_test(staff_required, login_url='index'), name='dispatch')
 class Usuario_Detail(View):
     def get(self, request, id):
         usuario = repository.get_by_id(id=id)
-        return render(request, 'usuarios/detail.html', {'usuario': usuario})
+        return render(request, 'usuarios/details.html', {'usuario': usuario})
 
 @method_decorator(user_passes_test(staff_required, login_url='index'), name='dispatch')
 class Usuario_Delete(View):
     def get(self, request, id):
         usuario = repository.get_by_id(id=id)
-        repository.delete(usuario)
+        if usuario:
+            repository.delete(usuario)
         return redirect('usuario_list')
 
 @method_decorator(user_passes_test(staff_required, login_url='index'), name='dispatch')
@@ -54,7 +53,7 @@ class Usuario_Update(View):
     def get(self, request, id):
         usuario = repository.get_by_id(id=id)
         user_form = self.form_class(instance=usuario)
-        return render(request, 'usuarios/update.html', {'user_form': user_form})
+        return render(request, 'usuarios/update.html', {'user_form': user_form, 'usuario': usuario})
 
     def post(self, request, id):
         usuario = repository.get_by_id(id=id)
@@ -62,4 +61,4 @@ class Usuario_Update(View):
         if user_form.is_valid():
             user_form.save()
             return redirect('usuario_detail', id=id)
-        return render(request, 'usuarios/update.html', {'user_form': user_form})
+        return render(request, 'usuarios/update.html', {'user_form': user_form, 'usuario': usuario})
